WIP (Renamed PlaylistPage and appropriate playlist names to CoursePage)
