+++ src/components/CoursePage/LessonItem.scss
