+++ src/components/CoursePage/LessonList.scss
