+++ src/components/CoursePage/LevelNavigation.scss
