+++ src/components/CoursePage/Playlist.scss
