+++ src/components/CoursePage/PlaylistNavigation.scss
