+++ src/constants/captions_nb.js
@@ -46,10 +46,9 @@ export default {
     },
     showhidefilter: 'Vis/skjul filter'
   },
-  playlist: {
+  coursepage: {
     courseinfo: 'Informasjon om kurset',
     courseinfonotfound: 'Oops, her har noen glemt å skrive kursinformasjon!',
-    instructionbutton: 'Lærerveiledning',
     levelnavigation: 'Hopp til',
     lessoncollections: 'Oppgavesamlinger',
     nomatchinglessons: 'Ingen oppgaver passer til filteret',
