+++ src/pages/CoursePage.scss
